Modify nsexec.c
