+++ nsexec/include/namespace.h
